% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_data.R
\name{wa_data}
\alias{wa_data}
\alias{get_wa_pu}
\alias{get_wa_locked_in}
\alias{get_wa_locked_out}
\alias{get_wa_features}
\alias{get_wa_attr}
\alias{get_wa_carbon}
\alias{wa_pu}
\alias{wa_locked_in}
\alias{wa_locked_out}
\alias{wa_weights}
\alias{wa_carbon}
\title{Washington data}
\format{
\describe{
\item{get_wa_pu}{\code{\link[terra:rast]{terra::rast()}} object.}
\item{get_wa_locked_in}{\code{\link[terra:rast]{terra::rast()}} object.}
\item{get_wa_locked_out}{\code{\link[terra:rast]{terra::rast()}} object.}
\item{get_wa_carbon}{\code{\link[terra:rast]{terra::rast()}} object.}
\item{get_wa_features}{\code{\link[terra:rast]{terra::rast()}} object.}
\item{get_wa_attr}{\code{\link[tibble:tibble]{tibble::tibble()}} object.}
}
}
\usage{
get_wa_pu()

get_wa_locked_in()

get_wa_locked_out()

get_wa_features()

get_wa_attr()

get_wa_carbon()
}
\description{
Conservation planning dataset for Washington, The United States of America.
}
\details{
The following functions are provided to import data:

\describe{

\item{\code{get_wa_pu()}}{Import planning unit data.
The planning units are a single layer \code{\link[terra:rast]{terra::rast()}} object.
Cell values denote land acquisition costs.
These data were originally obtained from Nolte (2020 a,b).}

\item{\code{get_wa_locked_in()}}{Import locked in data.
The locked in data are a single layer \code{\link[terra:rast]{terra::rast()}} object.
Cell values denote binary values indicating if each cell
is predominantly covered by protected areas
(excluding those with no mandate for biodiversity protection).
These data were originally obtained from USGS (2022)}

\item{\code{get_wa_locked_in()}}{Import locked out data.
The locked out data are a single layer \code{\link[terra:rast]{terra::rast()}} object.
Cell values denote binary values indicating if each cell
is predominantly covered by urban areas.
These data were originally obtained from the
Commission for Environmental Cooperation (2020)}

\item{\code{get_wa_carbon()}}{Import vulnerable carbon data.
The carbon data a single layer \code{\link[terra:rast]{terra::rast()}} object.
Cell values denote continuous values representing the amount
of carbon sequestered that is vulnerable to be released through
typical land-use conversion.
These data were originally obtained from the
Noon \emph{et al.} (2021, 2022)}

\item{\code{get_wa_features()}}{Import biodiversity feature data.
The feature data are a multi-layer
\code{\link[terra:rast]{terra::rast()}} object.
It contains the spatial distribution of 258 bird species.
To account for migratory patterns, data are provided for the
breeding and non-breeding distributions of species
(indicated by \code{"breeding"} and \code{"non-breeding"} in the layer names).
If a species is lacking such information,
then the species is denoted with its full distribution
(as indicated \code{"full"} in the layer names).
These data were originally obtained from the eBird Status and Trends
dataset (Fink \emph{et al.} 2020).}

\item{\code{get_wa_attr()}}{Import attribute data about the biodiversity features.
The feature attribute data are a data frame (\code{\link[tibble:tibble]{tibble::tibble()}}) object.
It contains taxonomic information for each feature
(i.e., layer in \code{get_wa_features()}) as well as estimates of
public interest (derived from Mittermeier \emph{et al.} 2021) and
extinction risk (based on the methodology of Davis \emph{et al.} 2018 and
and threat status classification data from IUCN 2025).
Since Mittermeier \emph{et al.} (2021)
did not contain public interest scores for all features,
scores were interpolated for features missing scores based on average
public interest score of features that belong to the same
taxonomic family. This object has the following columns:
\describe{
\item{feature}{Name of the feature (i.e., \verb{per get_wa_features()}).}
\item{binomial}{Taxonomic species and genus name of the feature.}
\item{family}{Taxonomic family name of the feature.}
\item{order}{Taxonomic order of the feature.}
\item{extinction_prob}{Probability of extinction.}
\item{interest_score}{Public interest score.}
}}
}
}
\examples{
# load packages
library(terra)

# import data
wa_pu <- get_wa_pu()
wa_features <- get_wa_features()
wa_attr <- get_wa_attr()
wa_locked_in <- get_wa_locked_in()
wa_locked_out <- get_wa_locked_out()
wa_carbon <- get_wa_carbon()

# preview planning units
print(wa_pu)
plot(wa_pu)
# preview locked in
print(wa_locked_in)
plot(wa_locked_in)

# preview locked out
print(wa_locked_out)
plot(wa_locked_out)

# preview features
print(wa_features)
plot(wa_features)

# preview attributes of features
print(wa_attr)

}
\references{
Commission for Environmental Cooperation. (2020). \emph{2015 Land Cover
of North America at 30 Meters}. North American Land Change Monitoring
System, 2nd Edition, \url{https://www.cec.org:443/north-american-environmental-atlas/land-cover-30m-2015-landsat-and-rapideye/}.

Davis M, Faurby S, and Svenning J-C (2018) Mammal diversity will take
millions of years to recover from the current biodiversity crisis.
\emph{Proceedings of the National Academy of Sciences}, 115: 11262--11267.

Fink D, Auer T, Johnston A, Ruiz-Gutierrez V, Hochachka WM and
Kelling S (2020) Modeling avian full annual cycle distribution and
population trends with citizen science data.
\emph{Ecological Applications}, 30: e02056.

IUCN (2025) The IUCN Red List of Threatened Species. Version 2025-2.
\url{https://www.iucnredlist.org}. Accessed on 13 May 2026.

Mittermeier JC, Roll U, Matthews TJ, Correia R, and Grenyer R (2021)
Birds that are more commonly encountered in the wild attract higher public
interest online. \emph{Conservation Science and Practice}, 3: e340.

Nolte C (2020a) \emph{Data for: High-resolution land value maps
reveal underestimation of conservation costs in the United States}.
Dryad, Dataset, \doi{10.5061/dryad.np5hqbzq9}.

Nolte C (2020b) High-resolution land value maps reveal
underestimation of conservation costs in the United States.
\emph{Proceedings of the National Academy of Sciences}, 117: 29577--29583.

Noon ML, Goldstein A, Ledezma JC, Roehrdanz PR, Cook-Patton SC,
Spawn-Lee SA, Wright TM, Gonzalez-Roglich M, Hole DG, Rockström J, and
Turner WR (2022) Mapping the irrecoverable carbon in Earth's ecosystems.
\emph{Nature Sustainability}, 5: 37--46.

Noon ML, Goldstein A, Ledezma JC, Roehrdanz PR, Cook-Patton SC,
Spawn-Lee SA, Wright TM, Gonzalez-Roglich M, Hole DG, Rockström J, and
Turner WR (2021) Mapping the irrecoverable carbon in Earth's ecosystems
(2.0) [Data set]. Zenodo. \doi{10.5281/zenodo.4091029}.

U.S. Geological Survey (USGS) Gap Analysis Project (GAP) (2022)
Protected Areas Database of the United States (PAD-US) 3.0: U.S. Geological
Survey data release, \doi{10.5066/P9Q9LQ4B}.
}
\keyword{datasets}
