% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prioritizrdata}
\alias{prioritizrdata}
\alias{prioritizr-package}
\title{prioritizrdata: Conservation Planning Datasets}
\description{
The \pkg{prioritizrdata} package is a supplemental package that contains
example datasets for conservation planning. It is intended to be used
alongside the \pkg{prioritizr} package.
}
\details{
This package contains the following datasets:

\describe{

\item{\link{tas_data}}{
Conservation planning dataset for Tasmania, Australia.
}

\item{\link{salt_data}}{
Conservation planning dataset for Salt Spring Island, Canada.
}

\item{\link{wa_data}}{
Conservation planning dataset for Washington, The United States of America.
}

}
}
