% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_data.R
\docType{data}
\name{salt_data}
\alias{salt_data}
\alias{salt_features}
\alias{salt_pu}
\alias{salt_pu}
\alias{salt_features}
\title{Saltspring Island conservation planning scenario}
\format{\describe{
  \item{salt_features}{\code{\link[raster]{RasterStack-class}} object}
  \item{salt_pu}{\code{\link[raster]{RasterLayer-class}} object.}
}}
\usage{
data(salt_features)

salt_pu

salt_features
}
\description{
This scenario involves making a prioritsation to adequately conserve
key ecological communities on Saltspring Island, British Columbia. For a worked example 
with this dataset, refer to the \code{{SaltSpring_Isl}} vignette. The 
scenario is intended to provide an example of how Raster-class input data can be used in 
\code{prioritizr}. The dataset contains the following data:
}
\details{
\describe{
  \item{\code{salt_pu}}{Planning unit data. A single band raster-class object 
  where each one hectare pixel contains the corresponding monetary cost of 
  acquisition, based on 2015 BC Land Assessment data.}
  \item{\code{salt_features}}{Biodiversity feature data. The probability of occurrence 
  of five key ecological communities found on Salt Spring island. Each layer in the stack 
  represents a different community type. In order these are; Old Forest, Savannah, 
  Wetland, Shrub, and a layer representing the inverse probability of occurrence of 
  human commensal species. For a given layer, the cell value indicates the composite probability of 
  encountering the suite of bird species most commonly associated with that community type.}
}
}
\examples{
# load data
data(salt_pu, salt_features)

# plot data
plot(salt_pu)
plot(salt_features)

}
\references{
This data was obtained as part of an online Marxan-based planning 
tool created for the Coastal Douglas-fir Conservation Partnership (CDFCP). For more information refer 
to \href{http://peter-arcese-lab.sites.olt.ubc.ca/files/2016/09/CDFCP_tutorial_2017_05.pdf}{Schuster et al. 2017}.
}
\keyword{datasets}
